+++ .ci/testruns/default/templates/testrun.yaml
@@ -24,16 +24,6 @@ spec:
         key: serviceaccount.json
 
   testflow:
-  - name: infrastructure-test
-    definition:
-      name: infrastructure-test
-      config:
-        - name: REGION
-          type: env
-          value: europe-west1
-        - name: RECONCILER
-          type: env
-          value: tf
   - name: infra-flow-test
     definition:
       name: infrastructure-test
@@ -44,16 +34,6 @@ spec:
         - name: RECONCILER
           type: env
           value: flow
-  - name: infra-migrate-test
-    definition:
-      name: infrastructure-test
-      config:
-      - name: REGION
-        type: env
-        value: europe-west1
-      - name: RECONCILER
-        type: env
-        value: migrate
   - name: bastion-test
     definition:
       name: bastion-test
