+++ pkg/controller/infrastructure/actuator.go
@@ -6,12 +6,18 @@ package infrastructure
 
 import (
 	"context"
+	"time"
 
 	"github.com/gardener/gardener/extensions/pkg/controller/infrastructure"
 	"github.com/gardener/gardener/extensions/pkg/terraformer"
+	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+	"github.com/go-logr/logr"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
+
+	"github.com/gardener/gardener-extension-provider-gcp/imagevector"
 )
 
 type actuator struct {
@@ -38,4 +44,29 @@ func CleanupTerraformerResources(ctx context.Context, tf terraformer.Terraformer
 		return err
 	}
 	return tf.RemoveTerraformerFinalizerFromConfig(ctx)
+}
+
+// NewTerraformer initializes a new Terraformer.
+func NewTerraformer(
+	logger logr.Logger,
+	restConfig *rest.Config,
+	purpose string,
+	infra *extensionsv1alpha1.Infrastructure,
+	disableProjectedTokenMount bool,
+) (
+	terraformer.Terraformer,
+	error,
+) {
+	tf, err := terraformer.NewForConfig(logger, restConfig, purpose, infra.Namespace, infra.Name, imagevector.TerraformerImage())
+	if err != nil {
+		return nil, err
+	}
+
+	owner := metav1.NewControllerRef(infra, extensionsv1alpha1.SchemeGroupVersion.WithKind(extensionsv1alpha1.InfrastructureResource))
+	return tf.
+		UseProjectedTokenMount(!disableProjectedTokenMount).
+		SetTerminationGracePeriodSeconds(630).
+		SetDeadlineCleaning(5 * time.Minute).
+		SetDeadlinePod(15 * time.Minute).
+		SetOwnerRef(owner), nil
 }
