+++ pkg/controller/infrastructure/actuator_delete.go
@@ -6,6 +6,7 @@ package infrastructure
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/gardener/gardener/extensions/pkg/controller"
 	"github.com/gardener/gardener/extensions/pkg/util"
@@ -13,11 +14,14 @@ import (
 	"github.com/go-logr/logr"
 
 	"github.com/gardener/gardener-extension-provider-gcp/pkg/apis/gcp/helper"
+	"github.com/gardener/gardener-extension-provider-gcp/pkg/controller/infrastructure/infraflow"
+	gcpinternal "github.com/gardener/gardener-extension-provider-gcp/pkg/gcp"
+	gcpclient "github.com/gardener/gardener-extension-provider-gcp/pkg/gcp/client"
 )
 
 // Delete implements infrastructure.Actuator.
 func (a *actuator) Delete(ctx context.Context, log logr.Logger, infra *extensionsv1alpha1.Infrastructure, cluster *controller.Cluster) error {
-	return util.DetermineError(a.delete(ctx, log, OnDelete, infra, cluster), helper.KnownCodes)
+	return util.DetermineError(a.delete(ctx, log, infra, cluster), helper.KnownCodes)
 }
 
 // ForceDelete forcefully deletes the Infrastructure.
@@ -25,21 +29,38 @@ func (a *actuator) ForceDelete(_ context.Context, _ logr.Logger, _ *extensionsv1
 	return nil
 }
 
-func (a *actuator) delete(ctx context.Context, log logr.Logger, selectorFn SelectorFunc, infra *extensionsv1alpha1.Infrastructure, cluster *controller.Cluster) error {
-	useFlow, err := selectorFn(infra, cluster)
+func (a *actuator) delete(ctx context.Context, log logr.Logger, infra *extensionsv1alpha1.Infrastructure, cluster *controller.Cluster) error {
+	infraState, err := a.infrastructureStateFromRaw(infra)
 	if err != nil {
 		return err
 	}
 
-	factory := ReconcilerFactoryImpl{
-		log: log,
-		a:   a,
+	credentialsConfig, err := gcpinternal.GetCredentialsConfigFromSecretReference(ctx, a.client, infra.Spec.SecretRef)
+	if err != nil {
+		return err
+	}
+	gc := gcpclient.New()
+	fctx, err := infraflow.NewFlowContext(ctx, infraflow.Opts{
+		Log:               log,
+		Infra:             infra,
+		Cluster:           cluster,
+		CredentialsConfig: credentialsConfig,
+		Factory:           gc,
+		Client:            a.client,
+		State:             infraState,
+	})
+	if err != nil {
+		return fmt.Errorf("failed to create flow context: %v", err)
 	}
 
-	reconciler, err := factory.Build(useFlow)
+	err = fctx.Delete(ctx)
 	if err != nil {
 		return err
 	}
 
-	return reconciler.Delete(ctx, infra, cluster)
+	tf, err := NewTerraformer(log, a.restConfig, "infra", infra, a.disableProjectedTokenMount)
+	if err != nil {
+		return err
+	}
+	return CleanupTerraformerResources(ctx, tf)
 }
