+++ pkg/controller/infrastructure/actuator_restore.go
@@ -17,24 +17,9 @@ import (
 
 // Restore implements infrastructure.Actuator.
 func (a *actuator) Restore(ctx context.Context, log logr.Logger, infra *extensionsv1alpha1.Infrastructure, cluster *controller.Cluster) error {
-	return util.DetermineError(a.restore(ctx, log, OnRestore, infra, cluster), helper.KnownCodes)
+	return util.DetermineError(a.restore(ctx, log, infra, cluster), helper.KnownCodes)
 }
 
-func (a *actuator) restore(ctx context.Context, logger logr.Logger, selectorFn SelectorFunc, infra *extensionsv1alpha1.Infrastructure, cluster *controller.Cluster) error {
-	useFlow, err := selectorFn(infra, cluster)
-	if err != nil {
-		return err
-	}
-
-	factory := ReconcilerFactoryImpl{
-		log:   logger,
-		a:     a,
-		infra: infra,
-	}
-
-	reconciler, err := factory.Build(useFlow)
-	if err != nil {
-		return err
-	}
-	return reconciler.Restore(ctx, infra, cluster)
+func (a *actuator) restore(ctx context.Context, log logr.Logger, infra *extensionsv1alpha1.Infrastructure, cluster *controller.Cluster) error {
+	return a.reconcile(ctx, log, infra, cluster)
 }
