+++ pkg/controller/infrastructure/infraflow/kubernetes_test.go
@@ -1,8 +1,8 @@
-// SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Gardener contributors
+// SPDX-FileCopyrightText: 2025 SAP SE or an SAP affiliate company and Gardener contributors
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package infrastructure
+package infraflow
 
 import (
 	"context"
