--- pkg/internal/infrastructure/templates/terraform.tfvars
@@ -1,1 +1,0 @@
-# New line is needed! Do not remove this comment.
