--- pkg/internal/infrastructure/templates/variables.tf
@@ -1,4 +1,0 @@
-variable "SERVICEACCOUNT" {
-  description = "ServiceAccount"
-  type        = string
-}
