+++ test/tm/apply_flow/maintain_test.go
@@ -16,8 +16,6 @@ import (
 	. "github.com/onsi/gomega"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	gcpinternal "github.com/gardener/gardener-extension-provider-gcp/pkg/gcp"
 )
 
 var shootName = flag.String("shoot-name", "", "name of the shoot")
@@ -29,12 +27,11 @@ func init() {
 var _ = Describe("Shoot maintenance testing", func() {
 	f := framework.NewGardenerFramework(nil)
 
-	framework.CIt("Testing if Shoot maintainance is successful with flow annotations", func(ctx context.Context) {
+	framework.CIt("Testing if Shoot maintenance is successful with flow annotations", func(ctx context.Context) {
 		shoot := &gardencorev1beta1.Shoot{ObjectMeta: metav1.ObjectMeta{Namespace: f.ProjectNamespace, Name: *shootName}}
 		Expect(f.GardenClient.Client().Get(ctx, client.ObjectKey{Namespace: f.ProjectNamespace, Name: *shootName}, shoot)).To(Succeed())
 
 		Expect(f.UpdateShoot(ctx, shoot, func(shoot *gardencorev1beta1.Shoot) error {
-			shoot.Annotations[gcpinternal.AnnotationKeyUseFlow] = "true"
 			shoot.Annotations[v1beta1constants.GardenerOperation] = v1beta1constants.ShootOperationMaintain
 			return nil
 		})).To(Succeed())
